requirements-fixed.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.1.3
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
et_xmlfile==2.0.0
fastapi==0.128.8
greenlet==3.3.1
h11==0.16.0
httptools==0.7.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
openpyxl==3.1.5
passlib==1.7.4
pillow==12.1.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
reportlab==4.4.9
SQLAlchemy==2.0.46
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
watchfiles==1.1.1
websockets==16.0
